Refactored openglrenderer.py, renderer2d.py, shape.py into Vispy2DRenderer.
