+++ p5/core/tests/test_svg.py
@@ -5,12 +5,8 @@ import xml.etree.ElementTree as etree
 from p5.core.svg import (
     get_style,
     parse_rect,
-    parse_circle,
-    parse_line,
-    parse_ellipse,
-    parser)
+    parse_line)
 from p5.core.color import Color
-from p5.core.shape import PShape
 
 svg = '<?xml version="1.0" encoding="utf-8"?> <!DOCTYPE svg PUBLIC "-//W3C//DTD SVG 20010904//EN" "http://www.w3.org/TR/2001/REC-SVG-20010904/DTD/svg10.dtd"> \
     <svg width="400px" height="300px" xml:lang="fr" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink"> \
