+++ p5/sketch/Vispy2DRenderer/__init__.py
@@ -1,0 +1,1 @@
+from .shape import *
