+++ p5/sketch/Vispy2DRenderer/openglrenderer.py
@@ -1,9 +1,9 @@
 from abc import ABC
 import numpy as np
 
-from ..core import p5
-from ..core.constants import SType, ROUND, MITER
-from ..core.primitives import Arc
+from p5.core import p5
+from p5.core.constants import SType, ROUND, MITER
+from p5.core.primitives import Arc
 
 from dataclasses import dataclass
 from vispy.gloo import Program, VertexBuffer, FrameBuffer, IndexBuffer
