+++ p5/sketch/Vispy2DRenderer/renderer2d.py
@@ -31,7 +31,7 @@ from vispy.gloo import VertexBuffer
 from contextlib import contextmanager
 from p5.sketch.shaders2d import src_texture
 from p5.sketch.shaders2d import src_line
-from p5.sketch.openglrenderer import OpenGLRenderer, get_render_primitives, COLOR_WHITE
+from p5.sketch.Vispy2DRenderer.openglrenderer import OpenGLRenderer, get_render_primitives, COLOR_WHITE
 
 
 class VispyRenderer2D(OpenGLRenderer):
