+++ p5/sketch/renderer3d.py
@@ -32,10 +32,10 @@ from contextlib import contextmanager
 
 from ..core.constants import Z_EPSILON
 from ..core.geometry import Geometry
-from ..core.shape import PShape
+from p5.sketch.Vispy2DRenderer.shape import PShape
 
 from ..pmath.matrix import translation_matrix
-from .openglrenderer import OpenGLRenderer, get_render_primitives, to_3x3, Style, COLOR_WHITE
+from p5.sketch.Vispy2DRenderer.openglrenderer import OpenGLRenderer, get_render_primitives, to_3x3, Style, COLOR_WHITE
 from .shaders3d import src_default, src_fbuffer, src_normal, src_phong
 from ..core.material import BasicMaterial, NormalMaterial, BlinnPhongMaterial
